// src/common/@generated/post/post-max-aggregate.input.ts
import { Field, InputType } from '@nestjs/graphql';

@InputType()
export class PostMaxAggregateInput {
  @Field(() => Boolean, {
    nullable: true
  })
  id?: true;

  @Field(() => Boolean, {
    nullable: true
  })
  title?: true;

  @Field(() => Boolean, {
    nullable: true
  })
  slug?: true;

  @Field(() => Boolean, {
    nullable: true
  })
  content?: true;

  @Field(() => Boolean, {
    nullable: true
  })
  published?: true;

  @Field(() => Boolean, {
    nullable: true
  })
  createdAt?: true;

  @Field(() => Boolean, {
    nullable: true
  })
  updatedAt?: true;

  @Field(() => Boolean, {
    nullable: true
  })
  authorId?: true;
}